# openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: info@librematch.org
    name: Libre:Match Team
    url: https://librematch.org
  description:
    "stable, reliable and feature-rich API for match information, leaderboards,\
    \ and tournament organisation for the Age of Empires multiplayer community"
  license:
    name: GNU AGPLv3
    url: https://www.gnu.org/licenses/agpl.txt
    x-last-modified: 1663947772699
  title: Libre:Match API
  version: 1.0.0-DRAFT
  x-logo:
    url: ""
servers:
  - description: Libre:Match Server v1
    url: https://api.librematch.org/v1
    x-last-modified: 1663947810574
security:
  - BasicAuth: []
  - X-Api-Key: []
paths:
...
  /teams:
    delete:
      description: ""
      operationId: remove_all_teams
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
        "501":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
          description: Not Implemented
          x-last-modified: 1665058119290
      summary: Remove all teams
      tags:
        - Teams
        - Team Administration
    get:
      operationId: get_team_collection
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
        "501":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
          description: Not Implemented
          x-last-modified: 1665058119290
      summary: Retrieve all available teams
      tags:
        - Information
        - Teams
        - Team Administration
    patch:
      description: ""
      operationId: partial_update_all_teams
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "415":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsupportedMediaTypeError"
          description: Unsupported Media Type
          x-last-modified: 1665057248987
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
        "501":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
          description: Not Implemented
          x-last-modified: 1665058119290
      summary: Partial update of all teams
      tags:
        - Administration
        - Team Administration
    post:
      description: ""
      operationId: create_new_team
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "415":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsupportedMediaTypeError"
          description: Unsupported Media Type
          x-last-modified: 1665057248987
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
        "501":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
          description: Not Implemented
          x-last-modified: 1665058119290
      summary: Create a new team
      tags:
        - Administration
        - Team Administration
    put:
      description: ""
      operationId: bulk_update_all_teams
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "415":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsupportedMediaTypeError"
          description: Unsupported Media Type
          x-last-modified: 1665057248987
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
        "501":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
          description: Not Implemented
          x-last-modified: 1665058119290
      summary: Bulk update of teams
      tags:
        - Administration
        - Team Administration
    summary: Endpoint for team information
    x-last-modified: 1665138931985
...
  /teams/{team_id}:
    get:
      operationId: get_details_for_team
      parameters:
        - deprecated: false
          description:
            The unique identifier (ULID) we use for teams that play on games
            using the Relic Link API
          explode: false
          in: path
          name: team_id
          required: true
          schema:
            $ref: "#/components/schemas/Ulid"
          style: simple
          x-last-modified: 1665069480649
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
        "501":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
          description: Not Implemented
          x-last-modified: 1665058119290
      summary: Retrieve details of a team
      tags:
        - Teams
        - Team Administration
    summary: Team information
    x-last-modified: 1665139023265
...
  /teams/{team_id}/stats:
    get:
      operationId: get_stat_collection_for_team
      parameters:
        - deprecated: false
          description:
            The unique identifier (ULID) we use for teams that play on games
            using the Relic Link API
          explode: false
          in: path
          name: team_id
          required: true
          schema:
            $ref: "#/components/schemas/Ulid"
          style: simple
          x-last-modified: 1665069480649
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
        "501":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
          description: Not Implemented
          x-last-modified: 1665058119290
      summary: Retrieve a list of stats for a specific team
      tags:
        - Statistics
        - Teams
        - Team Administration
    summary: Collection of stats for a team
    x-last-modified: 1665139090952
components:
...
  schemas:
...
    Ulid:
      additionalProperties: false
      properties:
        ulid:
          format: ulid
          maxLength: 26
          maximum: 26
          minimum: 26
          pattern: "^[0-7][0-9A-HJKMNP-TV-Z]{25}$"
          type: string
      required:
        - ulid
      type: object
      x-last-modified: 1665056674054
    UnsupportedMediaTypeError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 415
          minimum: 415
          title: code
          type: integer
      required:
        - code
        - message
      title: UnsupportedMediaTypeError
      type: object
      x-last-modified: 1665056674054
...
    NotImplementedError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 501
          minimum: 501
          title: code
          type: integer
      required:
        - code
        - message
      title: NotImplementedError
      type: object
      x-last-modified: 1665056674054
    TooManyRequestsError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 429
          minimum: 429
          title: code
          type: integer
      required:
        - code
        - message
      title: TooManyRequestsError
      type: object
      x-last-modified: 1665056674054
...
    NotAcceptableError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 406
          minimum: 406
          title: code
          type: integer
      required:
        - code
        - message
      title: NotAcceptableError
      type: object
      x-last-modified: 1665056674054
    NotFoundError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 404
          minimum: 404
          title: code
          type: integer
      required:
        - code
        - message
      title: NotFoundError
      type: object
      x-last-modified: 1665056674054
    ForbiddenError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 403
          minimum: 403
          title: code
          type: integer
      required:
        - code
        - message
      title: ForbiddenError
      type: object
      x-last-modified: 1665056674054
...
    UnauthorizedError:
      additionalProperties: false
      properties:
        message:
          maxLength: 32
          pattern: "^example-[0-9a-z]+$"
          title: Unauthorized access
          type: string
        code:
          maximum: 401
          minimum: 401
          title: code
          type: integer
      required:
        - code
        - message
      title: UnauthorizedError
      type: object
      x-last-modified: 1665056676293
...
  securitySchemes:
    BasicAuth:
      description: "Example: \n> Authorization: <Sample>"
      in: header
      name: BasicAuth
      type: apiKey
      x-last-modified: 1663947718456
    X-Api-Key:
      description: "Example: \n> X-Api-Key: ZGVtbzpwQDU1dzByZA=="
      in: header
      name: X-Api-Key
      type: apiKey
      x-last-modified: 1663947718456
    JWTAuth:
      bearerFormat: JWT
      description: "Example: \n> Authorization: Bearer <token>"
      scheme: bearer
      type: http
      x-last-modified: 1665071628782